agents/requirements.txt
deepagents
tavily-python
fastapi
python-dotenv
uvicorn
pydantic
psycopg[binary]
langgraph-checkpoint-postgres
